requirements.txt
streamlit>=1.29.0
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6